feat: Enhance transfer functionality and UI updates
- Added new API endpoints for completed user-to-user transfer records and user-to-team transfers.
- Updated HTML templates to reflect new routes and improved user experience for transfer history.
- Introduced a dropdown for selecting transfer type (user or team) in the transfer modal.
- Enhanced the display of transfer records, including completed transfers and rejection reasons.
- Improved breadcrumb navigation and button placements for better usability.
- Added pagination for transfer records and ensured proper balance displays in user accounts.

+++ Route/route_share_func_tools.go
@@ -43,7 +43,7 @@ func generateHTML(w http.ResponseWriter, template_data any, filenames ...string)
 		"mul": func(a, b int) int {
 			return a * b
 		},
-		"sub": func(a, b int) int {
+		"sub": func(a, b float64) float64 {
 			return a - b
 		},
 		"subtract": func(a, b int) int {
